requirements.txt
spacy==3.*
transformers>=4.40
pandas>=2.0
numpy>=1.25
scikit-learn>=1.3
regex>=2023.0.0
matplotlib>=3.7
PyYAML>=6.0
pdfminer.six>=20221105
